perf: impl. vectorized AsciiIgnoreCase.CommonPrefixLength and improve U8Scalar layout - Use it in AsciiIgnoreCase.Compare - Additional vector extensions for CommonPrefixLength and future AVX512 evaluation - Expose regular .CommonPrefixLength for convenience

+++ Sources/U8Scalar.cs
@@ -8,7 +8,7 @@ namespace U8Primitives;
 internal readonly struct U8Scalar
 {
     internal readonly byte B0, B1, B2, B3;
-    internal readonly byte Length;
+    internal readonly int Length;
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     internal U8Scalar(byte b)
